move configuration to its own package
